[sui-fork] Collapse RpcReader into ForkStore
RpcReader added a layer without owning any policy: its ObjectStore and
RuntimeObjectResolver impls forwarded to ForkStore's own, and the store
had already grown RPC-shaped helpers (latest_checkpoint_for_rpc,
chain_identifier, the index initializers) that existed only to serve it.
ForkStore now implements ReadStore, RpcStateReader, and RpcIndexes
itself, and startup hands the store to RpcService directly.

The routing table survives as a store.rs section with the per-method
rationale intact: fork-policy reads resolve through the local-first
helpers, no-policy surfaces (events, committees, layouts, ledger and
bitmap indexes) go straight to the stock reader, and the two hybrid
reads keep their stock-first fallback. Deleting the nine leaked helpers
removes every identical-signature inherent/trait collision, and moving
the impls into the store's module drops the pub(crate) surface the
split forced. Reader tests move to tests/store_rpc_traits.rs via the
existing #[path] pattern; the design docs now show RpcService and
Simulacrum standing directly on ForkStore.
